+++ internal/processing/ory/reconciliation_test.go
@@ -1,4 +1,4 @@
-package istio_test
+package ory_test
 
 import (
 	"context"
@@ -6,18 +6,25 @@ import (
 
 	gatewayv1beta1 "github.com/kyma-incubator/api-gateway/api/v1beta1"
 	. "github.com/kyma-incubator/api-gateway/internal/processing/internal/test"
-	"github.com/kyma-incubator/api-gateway/internal/processing/istio"
+	"github.com/kyma-incubator/api-gateway/internal/processing/ory"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	rulev1alpha1 "github.com/ory/oathkeeper-maester/api/v1alpha1"
+	"golang.org/x/exp/slices"
 	networkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 var _ = Describe("Reconciliation", func() {
-	When("multiple handlers in addtion to istio jwt", func() {
-		jwtConfigJSON := fmt.Sprintf(`{"authentications": [{"issuer": "%s", "jwksUri": "%s"}]}`, JwtIssuer, JwksUri)
+	When("multiple handlers in addtion to ory jwt", func() {
+		jwtConfigJSON := fmt.Sprintf(`
+		{
+			"trusted_issuers": ["%s"],
+			"jwks": [],
+			"required_scope": [%s]
+		}`, JwtIssuer, ToCSVList(ApiScopes))
+
 		jwt := []*gatewayv1beta1.Authenticator{
 			{
 				Handler: &gatewayv1beta1.Handler{
@@ -29,7 +36,7 @@ var _ = Describe("Reconciliation", func() {
 			},
 		}
 
-		It("with ory noop should provide istio vs and ory rule", func() {
+		It("with ory noop should provide only ory rules", func() {
 			// given
 			noop := []*gatewayv1beta1.Authenticator{
 				{
@@ -48,7 +55,7 @@ var _ = Describe("Reconciliation", func() {
 
 			// when
 			var createdObjects []client.Object
-			reconciliation := istio.NewIstioReconciliation(GetTestConfig())
+			reconciliation := ory.NewOryReconciliation(GetTestConfig())
 			for _, processor := range reconciliation.GetProcessors() {
 				results, err := processor.EvaluateReconciliation(context.TODO(), faceClient, apiRule)
 				Expect(err).To(BeNil())
@@ -71,16 +78,13 @@ var _ = Describe("Reconciliation", func() {
 
 				if vsOk {
 					vsCreated = true
-
 					Expect(vs).NotTo(BeNil())
-					Expect(len(vs.Spec.Http)).To(Equal(2))
-					Expect(len(vs.Spec.Http[1].Route)).To(Equal(1))
-					Expect(vs.Spec.Http[1].Route[0].Destination.Host).To(Equal(ServiceName + "." + ApiNamespace + ".svc.cluster.local"))
+					Expect(vs.Spec.Http[1].Route[0].Destination.Host).To(Equal(OathkeeperSvc))
 				} else if arOk {
 					oryRuleCreated = true
-
 					Expect(ar).NotTo(BeNil())
-					Expect(ar.Spec.Authenticators[0].Handler.Name).To(Equal("noop"))
+					expectedHandlers := []string{"jwt", "noop"}
+					Expect(slices.Contains(expectedHandlers, ar.Spec.Authenticators[0].Handler.Name)).To(BeTrue())
 					Expect(ar.Spec.Upstream.URL).To(Equal(fmt.Sprintf("http://%s.%s.svc.cluster.local:%d", ServiceName, ApiNamespace, ServicePort)))
 				}
 			}
@@ -88,7 +92,7 @@ var _ = Describe("Reconciliation", func() {
 			Expect(vsCreated && oryRuleCreated).To(BeTrue())
 		})
 
-		It("with ory oauth2 should provide istio vs and ory rule", func() {
+		It("with ory oauth2 should provide only ory rules", func() {
 			// given
 			oauthConfigJSON := fmt.Sprintf(`{"required_scope": [%s]}`, ToCSVList(ApiScopes))
 			oauth := []*gatewayv1beta1.Authenticator{
@@ -111,7 +115,7 @@ var _ = Describe("Reconciliation", func() {
 
 			// when
 			var createdObjects []client.Object
-			reconciliation := istio.NewIstioReconciliation(GetTestConfig())
+			reconciliation := ory.NewOryReconciliation(GetTestConfig())
 			for _, processor := range reconciliation.GetProcessors() {
 				results, err := processor.EvaluateReconciliation(context.TODO(), faceClient, apiRule)
 				Expect(err).To(BeNil())
@@ -134,16 +138,13 @@ var _ = Describe("Reconciliation", func() {
 
 				if vsOk {
 					vsCreated = true
-
 					Expect(vs).NotTo(BeNil())
-					Expect(len(vs.Spec.Http)).To(Equal(2))
-					Expect(len(vs.Spec.Http[1].Route)).To(Equal(1))
-					Expect(vs.Spec.Http[1].Route[0].Destination.Host).To(Equal(ServiceName + "." + ApiNamespace + ".svc.cluster.local"))
+					Expect(vs.Spec.Http[1].Route[0].Destination.Host).To(Equal(OathkeeperSvc))
 				} else if arOk {
 					oryRuleCreated = true
-
 					Expect(ar).NotTo(BeNil())
-					Expect(ar.Spec.Authenticators[0].Handler.Name).To(Equal("oauth2_introspection"))
+					expectedHandlers := []string{"jwt", "oauth2_introspection"}
+					Expect(slices.Contains(expectedHandlers, ar.Spec.Authenticators[0].Handler.Name)).To(BeTrue())
 					Expect(ar.Spec.Upstream.URL).To(Equal(fmt.Sprintf("http://%s.%s.svc.cluster.local:%d", ServiceName, ApiNamespace, ServicePort)))
 				}
 			}
